requirements.txt
cs50==9.2.3
Flask==3.0.0
Flask-Session==0.5.0
openai==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
